+++ src/shainet/inference/decoding.cr
@@ -1,5 +1,5 @@
 module SHAInet
-  private def self.softmax_with_temperature(logits : Array(Float64), temperature : Float64)
+  private def self.softmax_with_temperature(logits : Array(Float32), temperature : Float32)
     scaled = logits.map { |l| l / temperature }
     m = scaled.max
     exps = scaled.map { |l| Math.exp(l - m) }
@@ -7,12 +7,12 @@ module SHAInet
     exps.map { |e| e / sum }
   end
 
-  def self.top_k_sample(logits : Array(Float64), k : Int32, temperature : Float64 = 1.0, rng = Random::DEFAULT) : Int32
+  def self.top_k_sample(logits : Array(Float32), k : Int32, temperature : Float32 = 1.0, rng = Random::DEFAULT) : Int32
     raise ArgumentError.new("k must be > 0") unless k > 0
     raise ArgumentError.new("k cannot exceed logits size") if k > logits.size
 
     probs = softmax_with_temperature(logits, temperature)
-    pairs = [] of NamedTuple(index: Int32, prob: Float64)
+    pairs = [] of NamedTuple(index: Int32, prob: Float32)
     probs.each_with_index { |p, i| pairs << {index: i, prob: p} }
     top = pairs.sort_by { |t| -t[:prob] }.first(k)
     total = top.sum { |t| t[:prob] }
@@ -26,16 +26,16 @@ module SHAInet
     top.last[:index]
   end
 
-  def self.top_p_sample(logits : Array(Float64), p : Float64, temperature : Float64 = 1.0, rng = Random::DEFAULT) : Int32
+  def self.top_p_sample(logits : Array(Float32), p : Float32, temperature : Float32 = 1.0, rng = Random::DEFAULT) : Int32
     raise ArgumentError.new("p must be between 0 and 1") unless 0.0 <= p <= 1.0
 
     probs = softmax_with_temperature(logits, temperature)
-    pairs = [] of NamedTuple(index: Int32, prob: Float64)
+    pairs = [] of NamedTuple(index: Int32, prob: Float32)
     probs.each_with_index { |pr, i| pairs << {index: i, prob: pr} }
     sorted = pairs.sort_by { |t| -t[:prob] }
 
     cumulative = 0.0
-    chosen = [] of NamedTuple(index: Int32, prob: Float64)
+    chosen = [] of NamedTuple(index: Int32, prob: Float32)
     sorted.each do |t|
       cumulative += t[:prob]
       chosen << t
